Mixture Density Networks
(1) Added a mixture density function to CNet
(2) Whenever the forward operation is asked for, we compute the conditional mean of a mixture of gaussians.
(3) the network controls the means, variances and relative strenghts of these gaussians.
(4) This enables the network to be used on inverse problems with one-to-many type functions.