Column/Row Major issue fixed
(1) Mixture Density now a derived class of Discarnate layer, so that it does not need to be an output layer.
(2) While labview uses row-major formatting for its matrices, the eigen library uses column-major formats. Until now this was not a problem, since all inputs and outputs were quadratic and thus simply transposed.
It becomes an issue if one attempts to write a non-square  matrix to any of  the layer's weight matrices.